+++ src/Shopgate/Extended/ExtendedOrder.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopgate\Shopware\Shopgate\Extended;
+
+use ShopgateOrder;
+
+class ExtendedOrder extends ShopgateOrder
+{
+    use CloningTrait;
+    use CartUtilityTrait;
+
+    /**
+     * @param ShopgateOrder $order
+     * @return $this
+     */
+    public function loadFromShopgateOrder(ShopgateOrder $order): ExtendedOrder
+    {
+        $this->dataToEntity($order->toArray());
+        return $this;
+    }
+}